requirements.txt
Pillow==9.2.0
streamlit==1.22.0
streamlit-audiorec
streamlit-cropperjs